Move extension methods to Hosting.Abstractions - - Moved everything that could be moved to Hosting.Abstractions
